+++ Assets/com.mortise.compass/Runtime/Util/MathUtil.cs
@@ -7,8 +7,8 @@ namespace MortiseFrame.Compass {
 
         public static Node2D Pos2Node(Vector3 pos, int mpu, Vector2 localOffset, Map2D map) {
 
-            var x = Mathf.RoundToInt((pos.x - localOffset.x) * mpu - 1 / mpu / 2);
-            var y = Mathf.RoundToInt((pos.y - localOffset.y) * mpu - 1 / mpu / 2);
+            var x = Mathf.RoundToInt((pos.x - localOffset.x) * mpu - 1 / (float)mpu / 2);
+            var y = Mathf.RoundToInt((pos.y - localOffset.y) * mpu - 1 / (float)mpu / 2);
 
             x = Mathf.Clamp(x, 0, map.Width - 1);
             y = Mathf.Clamp(y, 0, map.Height - 1);
@@ -28,8 +28,8 @@ namespace MortiseFrame.Compass {
 
         public static Vector2Int Pos2Index(Vector3 pos, int mpu, Vector2 localOffset, Map2D map) {
 
-            var x = Mathf.RoundToInt((pos.x - localOffset.x) * mpu - 1 / mpu / 2);
-            var y = Mathf.RoundToInt((pos.y - localOffset.y) * mpu - 1 / mpu / 2);
+            var x = Mathf.RoundToInt((pos.x - localOffset.x) * mpu - 1 / (float)mpu / 2);
+            var y = Mathf.RoundToInt((pos.y - localOffset.y) * mpu - 1 / (float)mpu / 2);
 
             x = Mathf.Clamp(x, 0, map.Width - 1);
             y = Mathf.Clamp(y, 0, map.Height - 1);
